Return result_code from adb commands

+++ devices/adb.py
@@ -14,39 +14,44 @@ def adb_command(device, command, timeout = False):
     if timeout:
         timeout_adb_cmd = settings.TIMEOUT_CMD + " " + str(settings.EVAL_TIMEOUT) + " " + adb_cmd
         log_adb_command(device, timeout_adb_cmd)
-        os.system(timeout_adb_cmd + logger.redirect_string())
+        return os.system(timeout_adb_cmd + logger.redirect_string())
     else:
         log_adb_command(device, adb_cmd)
-        os.system(adb_cmd + logger.redirect_string())
+        return os.system(adb_cmd + logger.redirect_string())
 
 def shell_command(device, command, timeout = False):
-    adb_command(device, "shell " + command, timeout)
-    #os.system("$ANDROID_HOME/platform-tools/adb -s " + device + " shell " + command)
+    return adb_command(device, "shell " + command, timeout)
 
 def sudo_shell_command(device, command, timeout = False):
-    shell_command(device, "\" su -s sh -c '" + command + "'\"", timeout)
-    # os.system("$ANDROID_HOME/platform-tools/adb -s " + device + " shell \" su -c '" + command + "'\"")
+    return shell_command(device, "\" su -s sh -c '" + command + "'\"", timeout)
 
 def push(device, src, dest):
-    adb_command(device, "push " + src + " " + dest)
+    return adb_command(device, "push " + src + " " + dest)
 
 def sudo_push(device, src, dest):
     filename = os.path.basename(src)
     aux_file = "/sdcard/" + filename
-    push(device, src, aux_file)
-    sudo_shell_command(device, "cat " + aux_file + " > " + dest)
-    sudo_shell_command(device, "rm " + aux_file)
+
+    result_code = push(device, src, aux_file)
+    if result_code != 0: raise Exception("Unable to sudo_push file: " + src)
+
+    result_code = sudo_shell_command(device, "cat " + aux_file + " > " + dest)
+    if result_code != 0: raise Exception("Unable to sudo_push file: " + src)
+
+    result_code = sudo_shell_command(device, "rm " + aux_file)
+    if result_code != 0: raise Exception("Unable to sudo_push file: " + src)
+
     return filename
 
 def pull(device, src, dest):
-    adb_command(device, "pull " + src + " " + dest)
-
+    return adb_command(device, "pull " + src + " " + dest)
 
 def uninstall(device, package_name):
-    adb_command(device, "uninstall " + package_name)
+    return adb_command(device, "uninstall " + package_name)
 
 def install(device, package_name, apk_path):
-    adb_command(device, "install " + apk_path)
+    result_code = adb_command(device, "install " + apk_path)
+    if result_code != 0: raise Exception("Unable to install apk: " + apk_path)
 
     cmd = adb_cmd_prefix + " -s " + device + " shell pm list packages | grep " + package_name
     log_adb_command(device, cmd)
@@ -60,7 +65,7 @@ def pkill(device, string):
     pkill_cmd = adb_cmd + "ps | grep " + string + " | awk '{print $2}' | xargs -I pid " + adb_cmd + "kill pid "
     log_adb_command(device, pkill_cmd)
 
-    os.system(pkill_cmd + logger.redirect_string())
+    return os.system(pkill_cmd + logger.redirect_string())
 
 def get_battery_level(device):
     adb_cmd = adb_cmd_prefix + " -s " + device + " shell "
